sync
This improves permission handling
now only admins or creators can edit/delete their investigations

+++ core/web/frontend/generic.py
@@ -10,7 +10,7 @@ from core.errors import GenericValidationError
 from core.indicators import Regex
 from core.database import AttachedFile
 from core.web.helpers import get_object_or_404
-from core.web.helpers import requires_permissions, group_user_permission
+from core.web.helpers import requires_permissions, group_user_permission, get_user_groups
 
 binding_object_classes = {
     "malware": Malware,
@@ -34,7 +34,7 @@ class GenericView(FlaskView):
     def get(self, id):
         obj = self.klass.objects.get(id=id)
         if hasattr(obj, "sharing"):
-            if group_user_permission():
+            if group_user_permission(obj):
                 return render_template(
                     "{}/single.html".format(self.klass.__name__.lower()), obj=obj)
             else:
@@ -75,21 +75,30 @@ class GenericView(FlaskView):
     @requires_permissions("write")
     @route('/edit/<string:id>', methods=["GET", "POST"])
     def edit(self, id):
+        obj = self.klass.objects.get(id=id)
+        #ToDo Group admins support
+        if current_user.username != obj.created_by or not current_user.has_role('admin'):
+            return redirect(request.referrer)
+
         if request.method == "POST":
             return self.handle_form(id=id)
-        obj = self.klass.objects.get(id=id)
+
         form_class = obj.__class__.get_form()
         form = form_class(obj=obj)
         return render_template(
             "{}/edit.html".format(self.klass.__name__.lower()),
             form=form,
             obj_type=self.klass.__name__,
-            obj=obj)
+            obj=obj,
+            groups=get_user_groups())
 
     @requires_permissions("write")
     @route('/delete/<string:id>', methods=["GET"])
     def delete(self, id):
         obj = self.klass.objects.get(id=id)
+        #ToDo Group admins support
+        if current_user.username != obj.created_by or not current_user.has_role('admin'):
+            return redirect(request.referrer)
         obj.delete()
         return redirect(
             url_for('frontend.{}:index'.format(self.__class__.__name__)))
@@ -108,9 +117,11 @@ class GenericView(FlaskView):
             obj = self.klass.objects.get(id=id)
             klass = obj.__class__
             form = klass.get_form()(request.form, initial=obj._data)
-
+        #import code; code.interact(local=dict(globals(), **locals()))
         if form.validate():
             form.populate_obj(obj)
+            if form.formdata.get("sharing") and hasattr(klass, "sharing_permissions"):
+                obj.sharing_permissions(form.formdata["sharing"], obj.sharing)
             try:
                 self.pre_validate(obj, request)
                 obj = obj.save(validate=not skip_validation)
@@ -122,7 +133,8 @@ class GenericView(FlaskView):
                     "{}/edit.html".format(self.klass.__name__.lower()),
                     form=form,
                     obj_type=klass.__name__,
-                    obj=None)
+                    obj=None,
+                    groups=get_user_groups())
             except NotUniqueError as e:
                 form.errors['Duplicate'] = [
                     'Entity "{}" is already in the database'.format(obj)
@@ -131,7 +143,8 @@ class GenericView(FlaskView):
                     "{}/edit.html".format(self.klass.__name__.lower()),
                     form=form,
                     obj_type=klass.__name__,
-                    obj=None)
+                    obj=None,
+                    groups=get_user_groups())
 
             # success - redirect to view page
             return redirect(
